Add files via upload
This commit includes Bag of Words and TF-IDF conversion from words to vectors and then after converting it uses ML Model to train the data. It also uses stop words to remove common words and apart from that it uses lemmatizer and stemmer to get the root word from the word.